# Replace Regexp with split or concat if possible